perf: 레이싱 화면 내 리렌더링 최소화
