+++ packages/user/src/components/event/racing/controls/ControlButton.tsx
@@ -1,5 +1,6 @@
 import type { Category } from '@softeer/common/types';
-import { useCallback, useEffect, useRef, useState } from 'react';
+import numeral from 'numeral';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import useAuth from 'src/hooks/useAuth.tsx';
 import { useToast } from 'src/hooks/useToast.ts';
 import type { Rank } from 'src/types/racing.d.ts';
@@ -28,18 +29,23 @@ export default function ControlButton({
 	type,
 	data,
 }: ControlButtonProps) {
-	const { rank, percentage } = data;
+	const { rank, vote, percentage } = data;
 	const { progress, handleClick } = useGaugeProgress({
 		percentage,
 		onCharge,
 		onFullyCharged,
 	});
 
+	const displayVoteStats = useMemo(
+		() => `${percentage.toFixed(1)}% (${formatVoteCount(vote)})`,
+		[percentage, vote],
+	);
+
 	return (
 		<ControllButtonWrapper rank={rank}>
 			<Gauge percent={progress} />
 			<ChargeButtonWrapper onClick={handleClick} type={type}>
-				<ChargeButtonContent type={type} {...data} />
+				<ChargeButtonContent type={type} rank={rank}>{displayVoteStats}</ChargeButtonContent>
 			</ChargeButtonWrapper>
 		</ControllButtonWrapper>
 	);
@@ -78,4 +84,22 @@ function useGaugeProgress({
 	}, [onCharge, onFullyCharged, isAuthenticated, percentage, toast]);
 
 	return { progress, handleClick };
+}
+
+/** Utility Functions */
+function formatVoteCount(count: number): string {
+	const formatted = numeral(count).format('0,0'); // 기본 포맷팅
+	return convertToKoreanUnit(formatted);
+}
+
+function convertToKoreanUnit(formatted: string): string {
+	const number = parseFloat(formatted.replace(/,/g, ''));
+
+	if (number >= 100000000) {
+		return `${(number / 100000000).toFixed(2)}억`;
+	}
+	if (number >= 10000) {
+		return `${(number / 10000).toFixed(2)}만`;
+	}
+	return formatted;
 }
